+++ src/cli.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python3
 
-from api import Interface
+from turing_chat import TuringChat
 from translate import tr
 
 CURSOR_UP_ONE = '\x1b[1A'
@@ -11,9 +11,6 @@ interface = None
 
 def printMessage(text, message=False, username=None):
     global interface
-    if username:
-        interface.otherUsername = username
-        return printMessage(tr("username.other.changed") + username)
     if message:
         username = interface.otherUsername
         if(not username):
@@ -23,7 +20,7 @@ def printMessage(text, message=False, username=None):
         print("\r" + ERASE_LINE + text + "\n" + tr("command.label") + " : ", end="")
 
 
-interface = Interface(printMessage)
+interface = TuringChat(printMessage)
 
 while True:
     try: